Add Advantage2 keyboard and new bindings for it
After having bought a Kinesis Advantage2 keyboard, it became quite
obvious to also change to the US qwerty layout. After this decision and
due to the new layout of the keys, I added karabiner bindings for the German
umlauts and for the special function previously activated by function and
one of the function keys. Things like media control, volume control and
display brightness control are now set to ctrl + f<n>.