Add keybindings for moving code without Esc+
Keybindings like <A-k> and <M-k> require the left option key to work as
a meta/alt key, which is not the case in macOS by default. Instead by
default you can use the left option key to create symbols.

To use the left/right option keys as meta keys in iTerm, one has to set
the mode to Esc+ (see Preferences > Profiles > Keys). Then the
keybindings using <A- and <M- would work as expected, but... no symbols
possible anymore and in my case also the karabiner macros for diacritics
wouldn't work anymore due to the escape sequences.

Therefore I now use mappings with the symbols that the left options key
would produce, i.e. ∆ for <A-j>. Though, those "used symbols" are not
available for typing in text anymore.