Remove hyper key definition for now
The hyper key definition interferes with my workflow in Vim because the
escape use case has a delay of 1000 milliseconds before a click on caps
lock is considered an escape. I'm often faster and then carry out a line
movement instead of exiting insert mode. Adjusting the delay timer would
have been one way to solve this issue but as I currently do not use the
hyper key (yet), I simpl drop it for the moment.