Add global theme switcher command
First the profile now defines a environment variable GLOBAL_THEME that
gets exported when shell is started. Based on this variable tmux will
load the respective theme file.
Second a new shell script is now available to switch theme on the fly,
which means that the tmux theme is changed if tmux is running and
additionally an escape sequence is sent to iTerm2 to also change its
color palette for the current window/pane by checking if running in tmux
or not and sending the correct sequence.