Switch font to JetBrainsMono and separate Symbols
As iTerm2 and others (Kitty, Wezterm, ...) are able to deal with more
than one font, it's no longer needed to use patched/broken font for the
symbols. In addition I like to try a new font that's more subtle and
supports ligatures. JetBrains Mono is the chosen one together with the
Symbols only nerd font.