feat(wezterm): first sketch of a status bar
Now this is a replica of the right part of my tmux status bar. I'm
pretty sure I will leave tmux behind once wezterm has the amenities I'm
used to.

Still missing is the design of the left part and overall increase of the
font size in the status area.