fix(nvim): clear formatexpr to make gq* function work again
With LSP formatexpr gets set to a LSP function that tries to do the
right thing depending on the fileype. While this sound fantastic in
theory this does not work for markdown for me and things like `gqq` and
`gqap` break, which use a lot in markdown files. Therefore this change
clears the `formatexpr` option for markdown files.